requirements.txt
pytest~=8.3.4
allure-python-commons~=2.13.5
selenium~=4.27.1